loader: Do not close fds automatically on exit
We are storing global references to file descriptors, so we cannot release
them, as they may have already been released by the task itself = silent double
free! This messes up forked tasks references.

For now we are going to rely on user tasks to close file descriptors until we
find a better solution.